move Asimov data generation to model utils
